refactor: renamed ColAutoGrow to just AutoGrow
