+++ src/components/Col/Col.stories.js
@@ -5,7 +5,7 @@ import { storiesOf } from '@storybook/react';
 import { withKnobs } from '@storybook/addon-knobs';
 import withReadme from 'storybook-readme/with-readme';
 
-import { Col, Fluid, Responsive, ColAutoGrow } from './stories';
+import { Col, Fluid, Responsive, AutoGrow } from './stories';
 import README from './README.md';
 
 storiesOf('Col', module)
@@ -14,4 +14,4 @@ storiesOf('Col', module)
   .add('Default', () => <Col />)
   .add('Fluid example', () => <Fluid />)
   .add('Responsive example', () => <Responsive />)
-  .add('ColAutoGrow example', () => <ColAutoGrow />);
+  .add('Auto grow example', () => <AutoGrow />);
