+++ src/components/Col/stories/index.js
@@ -1,4 +1,4 @@
 export Col from './Col';
 export Fluid from './Fluid';
 export Responsive from './Responsive';
-export ColAutoGrow from './ColAutoGrow';
+export AutoGrow from './AutoGrow';
